Champion shuffling, level & preloading all working
Now I need to code all the main game logic. Ability to click on a champ
card, and keeping the game alive for the whole round until you click
all the champions that you had not clicked on yet. Once you get to
the end of a round, then go to the next level.

Also if you ever click on the same champion inside of a round,
then Game Over.

Not sure how I'll do this, but it should be achievable with a few
simple state variables. I'm also thinking of adding another array that
keeps track of which champions you have selected already, or which ones
you have left to select.
For testing this, it would be useful to display this array on screen.
I think shuffling after picking should already be taken care of,
since it shuffles on every render. Not sure tho, bc the new state
variable that keeps track of selection, wont be displayed on the screen.
So I dont see why the components would re-render even if I update this
state variable. Not sure