Basic working Game with Cheat Logs & onScreen
Need to add 'Score'. I have to look back at the example to see how
'score' should really work

Most importantly:
I have to think about how I want to handle Game Over, in terms of the
interface. Right now it just simply restarts. But I find that as a
player, I like to verify that I did in fact click on a card 2 times
in a single round. So right now, I look at the console, and the
champions selected cheat log that I render on the screen, to verify.
So I think I should also give the player some way of verifying by
themselves. I can think of 3 options
1. Add Game Over Screen, using a modal. Include a text area that
displays the logs in order, or the selected champs in order + the
champs you had yet to click + the one you clicked twice
2. Add this screen, or just the functionalities, but without using
a modal. Probably with some buttons & on top of the Board. At the top
is best, because the board expands downwards.
3. leave cheat logs as a button toggle

Also, I might not want the board to dissapear or reset to
lvl 1 immediately, because this might make it harder for the player
to 'verify'. Have to think about all this

Also, heres a concern I wrote in the previous commit, that is worth
revisiting & thinking about:
I think shuffling after picking should already be taken care of,
since it shuffles on every render. Not sure tho, bc the new state
variable that keeps track of selection, wont be displayed on the screen.
So I dont see why the components would re-render even if I update this
state variable. Not sure