Added some card themes. Cleaned up Visual bugs
For card themes, now we have Playing Cards (default), Genshin,
and League. Currently the only problem im running into, is that
when I advance levels, the new images for Playing Cards and Genshin
dont seem to have been pre-loaded. Not sure if there is a solution for
this. The League cards are laoded from urls, and preload perfectly,
but Playing Cards and Genshin load from static files, and they dont
seem to preload correctly on new renders.