Fixed 'Best Score bug'. And more minor adjustments
This is right before replacing cardsThisLevel state variable with a
useMemo variable. Will be changing some of the useEffects to fit this.

The Best score bug was basically our logic for calculating best score
was flawed, because if we cleared a level, we were adding 1 to our
best score. This works fine if you clear every level, but if you clear
for example up to level 3, and then fail level 4 and reset to level 1,
then upon clearing level 1 and going to level 2, your previous
best Score of 8 would go up to a Best Score of 9, instead of not
changing like its suppossed to.