Added cardsClickedCheat (visual indicator)
Also cleaned up a bit & minor adjustemnts/fixes.

TODO TODO BIG.
I now think that the way that I have a separate array for cardsClicked
is maybe wrong. The problem with this, is that now in order to make
this cardsClickedCheat work, I have to have each card map over
cardsClicked, in order to check if it this card has been clicked before.

I think instead of this, it would be much more useful to just have each
card/character object have a property called 'clicked'. This property
would be set to false on each new level, and when a card is clicked,
it would be set to true. This way, each Card would easily know if itself
has been clicked very easily by checking this own propery, instead of
having each Card map through the cardsClicked array.
I would also be saving a state varaiable, but I have to think about this
further...
I remember that I have a useEffect that runs anytimes cardsClicked
changes.. so I have to look back and really think about this & how the
change would look, before proceeding