Removed cardsClicked as a state variable.
Due to this, I also had to make score a state variable, which makes
sense.
Basically, I realized that I dont need this separate array state
variable where I keep adding each card that is clicked every level,
and then clearing it when it needs to be cleared. Instead, I can simply
create an object property on each 'card' in the orginal array, and this
property is called 'isClicked'. This propery on each card will keep
track of whether the card has been clicked on the current level.
If I try to compare my game to other simple games I have made before,
this approach makes perfect sense- having each object have a property to
keep track of its state.
This also helps to avoid the loop checking between 2 arrays that I
was perviously forced to do, in order to display the cardsClickedCheat
in red.
Previously I was making EACH card on the board run cardsClicked.some()
to check if it was also on the list of cardsClicked. This nested looping
seemed super suboptimal, so its one of the reasons I made this change.