Fixed error and loading states & their display
Also in the App.js JSX, I am now simply passing error and loading state
to the <Board/> component. The Board component is responsible for
checking and displaying either the error, the loading screen, or if
there is no error and loading finished, then it displays the board

TODO TODO
I think i will now put the choosing card theme useEffect into 1
custom Hook. This will extract this big logic and some of the code
from the main app, for example, the loading and error states.

This hook will have to return a lot of things though to work proerly.
The cardsList, setCardsList (for shuffling), error, isLoading. Also
it will have to take in the parameter 'cardTheme'. Not sure though if
extracting this logic is actually useful/prductive or not, but I think
its in the right direction, so I will try it.

Originally, Ive been wondering if each of the 3 fetches: 'playingCards',
'league', 'genshin', should all be put into their own useEffect.
However, this really complicates things, because if each was in their
own useEffect, then I couldnt put them inside of the switch statement
that they are currently on. So this makes me think that it is ok and
that there is no reason to put each in its own custom Effect, and
instead I should keep them as pure async fetch call functions.