Extracted cardTheme loading into its custom Hook
This new custom hook 'useCards' takes care of the state when setting
a new card theme. It also takes care of the loading, error, and
image pre-loading.

The only thing that I'm wondering about, is whether I should keep
the image pre-loading inside of this custom hook, or if I should move
it to a useEffect ouside on the parent caller. The thing is, that
right now, I like how the 'loading' state waits for all the images
to be pre-loaded to THEN set the 'loading' state to false. But If I
extracted the pre-loading function outside of the custom hook 'useCards'
then I wouldnt be able to do this.
Not sure, but I think the current version is best.

Other possible improvement is to move the top function shuffleArr into
a helpers.js file. This is because we use shuffleArr in both useCards
and also App.js, however since it is a very short function, we are
currently just re-writting it. The thing is that it seems overkill to
create an entire helpers.js just for this one function.... not sure..
but yea I should prob extarct it into a helpers.js tbh