Improved context use, logic, styles, localStorage
Made the theme toggle switch smaller, and adjusted color + off icon.
Also added border to the slider.
Moved context to its own file, and used theme context for the footer
as well to fix its visibility.
Also now use localStorage to persist theme

TODO TODO
I think it looks more finalized now, dont think theres anything left
to add to our use of context.
Next thing to add to the app would probably be changing the board theme.
The problem is that by this point, adding any extra options to the game
kind of saturates the screen with too many options & inputs.

Board theme would be a drop-down with 2-3 preset themes. Maybe a green
background color, and one with an wood image background.
Could also add the ability to upload any image and set that as the
background, or give an image url.
But this presents 2 main extra problems:
1. Do I want to move a lot of options to an html form element ?
2. Do I change the current looks of the boards + cards, so that cards
no longer have an individual border, or so that the board has a fixed
size?

Choosing a different sizing solution for the board + cards is a big
problem by itself, but this would be the only way that 'image uplaod
backgrounds' would make any sense bc otherwise the image wouldnt look
good with the current changing number of cards changing the board size.

TODO TODO
Overall, I do think that the next step is probably to make the board
always a rectangle (color the board background instead of each
individual card background). Also, remove each card border.
Have to think about all of this a lot tho, and decide how we want it to
look next.
As far as the html form element containing the options, I find this
to be suboptimal, because it would add extra clicks when opening and
closing the form modal, to see the option changes take place.
Right now you can easily change between all 3 card themes, and toggle
show names, and instantly see the results, without having to do a lot
of extra clicking to show & close a form modal.