created & moved into Theme and Log Contexts
Also moved <SettingsModal/> into <GameOptions/> to simplify the logic.
Initially, I wanted to have SettingsModal as a child of App.js, however
this complicated the logic for opening the modal. This is because the
only solutions I found to trigger the opening of the modal was either
leaving all the modal logic in App.js, which seems wrong/ bloated.
Or to pass a prop to SettingsModal, and then inside of it run
an effect that checks if (openProp===true) runOpeningCode().
The problem with this is that if you open the modal, then move
your mouse really fast, the animation skips/breaks for a tiny moment.
Im guessing this is just because of how useEffect runs after the state
render, and isnt insant.
So to simplify, I put the modal inside of GameOptions. I could maybe
still try the last approach I mentioned, but instead make
use of the 'open' prop ON THE DIALOG element itself. The reason why
I didnt try this is bc I previousely played around with the open prop,
and it complicated the display/state of the modal.
It seemed like it was only there for type"dialog" and not tyoe="modal".
However, this was back when I was having problems with animation + state
when closing the modal, so maybe now I can revisit some variation of
using the 'open' prop on the dialog itself

Also have to think about whether i want to create a game context, and
create <GameView>, and extract some of the main game logic into a
useGame() which would be used inside of <GameView/>.

But for now, the plan is to work on the look and feel of the app, a few
look/UI design decisions, and maybe a few organizational things.
TODO TODO refer to last commit.