Added CardBack support + made X an svg to simplify
X that shows on gameOver, will now always draw from corner to corner
according to width and height values. I still have 2 sets of w and h
values: one pair in css and another in JSX. It would be nice to have
only one place, but because of how the SVG needs these numbers to draw,
I prob cant achieve this unless I use a CSS-in-JS styling solution, not
sure

CardBack support works fine & uses localStorage. Think the way it works
could be improved, but Im not sure how... I wanted to avoid writting
both a mapping object AND a naming(?) object.
But if I dont do that, then I feel like there is no wayt to turn the map
into a set of objects... which would be handy for things like a
name property. So I shoudl think about this, and what the solution
would looks like if I have 2 objects, one to hold the map.VARIABLES,
and the second object would be the one with all the properties like
{name, fileName, url}. Not sure if I even need to add these properties
tho... I feel like I dont, unless I need diff names than the key names
in the map, such as if I dont want to use 'YUGIOH' as name

TODO refer to prev commits