Cleaned up a lil from GameContext implementation
TODO TODO, figuire out if we can change how the flipping animation works
or if it would be worth it to go back to not using useContext.

Right now GameContext renders 6 times on startup, and 3 on every card
click. I found that my flipping animation is the reason for this, which
makes sense bc of the 3 setState calls separated by time, but it was
unexpected.

Think about whether this a problem to begin with, or if its ok. If its
not ok, think about whether we should try to not have a GameContext
again. If we go this route tho, we should start from here, and move
forward instead of reverting to the prev commit before I implemented.
This is because when I went this route, I ended up finding & solving a
lot of small organization/architect and logical rendering/object keys
problems that coulld be solved.

I think though that this extra re-renderring might be ok because of the
delay. Something we can do also for example, is try and useMemo or
useCallback on expensive things from child components that consume
the GameContext. Im mainly thinking of the logic in Card.js that finds
if cards are from 'playingCards', and then shorten names and adds icons.
This logic needs to either be moved into the useCards or
fetchPlayihngCards, or it needs to be useMemo'd or Something

Anwyays, find out if animation can be done differently too, to reduce
the renders