Fixed few excess functions using useMemo, Callback
In Card.js, when it shortens names for playingCards, and adds the
suit icon to the name.
Also in SettingsModal.js, when it displays the different cardBacks

These functions were previousely running 3 times each, since the
components have to run 3 times, because of how the flipping animation
works in GameContext.

TODO TODO
Im thinking that I can probably change the flipping animtion and
separate it. Right now its in an effect, so that it changes on cardsList
however maybe there is a better way to do this without putting it in
an effect. Regardless.. I need to think about this an whether I can
revamp the way the flipping animation works, so that gameConetxt doesnt
render 3 times and forces everything under it to render 3 times as well.