Changed deckList fetch from mutation to query
-Initially, I had made this a query but ran into many problems, Some
stemmed from bad naming that was then causing me to write bad logic.
One I moved it to a mutation and cleaned up a lot of the naming,
I now moved it back to a query and have no bugs/problems

-Also I now decided to just let our component change state.cardTheme,
and have our query simply react to the cardTheme change. This is ok

-staleTime set to 60s per query

- In our apiSlice, I decided to make the preLoading images run after
our queryFn fetch completes. This makes the preLoading of images
run in parallel with the useEffect described above (the useEffect runs
handleNewDeck).
The reason for this is that if I instead include the preLoadImages
inside of the queryFn, then it adds a small but noticable delay before
the cards switch to the new ones (maybe 0.5 to 1s)
Also, I dont want this preLoading logic to be insde Board.js or any
React component, because it bloats the component with unrelated code

TODO
- clean up / delete some comments
- There is currently a useEffect that runs when our query data changes,
in of Board.js. It just runs handleNewDeck(data) anytime my Query
returns new data. I need this to run regardless of whether the data is
retrned from the cache or from a fetch, and this was the most
straightforward approach. I am not sure if maybe there is a better way
to code this without the useEffect/outside React components.
- maybe conside in the future how preLoadingImages logc & code placement
can be improved